# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dehasoft E-Ticaret Çözümü",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/seller/offer/product/add": {
            "post": {
                "tags": [
                    "Kampanya"
                ],
                "summary": "Ürüne Kampanya Tanımla",
                "description": "Var olan ürüne var olan bir kampanya tanımlaması yapılır.",
                "operationId": "offerProductAdd",
                "requestBody": {
                    "description": "Var olan ürüne var olan bir kampanya tanımlaması yapılır.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "discount_id"
                                ],
                                "properties": {
                                    "discount_id": {
                                        "type": "integer",
                                        "example": "Kampanya Id"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": "Ürün Id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kampanya tanımlandı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kampanya tanımlandı."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "deha_token": []
                    }
                ]
            }
        },
        "/api/seller/offer/category/add": {
            "post": {
                "tags": [
                    "Kampanya"
                ],
                "summary": "Kategoriye Kampanya Tanımla",
                "description": "Var olan kategoriye var olan bir kampanya tanımlaması yapılır.",
                "operationId": "offerCategoryAdd",
                "requestBody": {
                    "description": "Var olan kategoriye var olan bir kampanya tanımlaması yapılır.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "discount_id"
                                ],
                                "properties": {
                                    "discount_id": {
                                        "type": "integer",
                                        "example": "Kampanya Id"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "Kategori Id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kampanya tanımlandı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kampanya tanımlandı."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "deha_token": []
                    }
                ]
            }
        },
        "/api/seller/offer/user/add": {
            "post": {
                "tags": [
                    "Kampanya"
                ],
                "summary": "Kullanıcıya Kampanya Tanımla",
                "description": "Var olan kullanıcıya var olan bir kampanya tanımlaması yapılır.",
                "operationId": "offerUserAdd",
                "requestBody": {
                    "description": "Var olan kullanıcıya var olan bir kampanya tanımlaması yapılır.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "discount_id"
                                ],
                                "properties": {
                                    "discount_id": {
                                        "type": "integer",
                                        "example": "Kampanya Id"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": "Kullanıcı Id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kampanya tanımlandı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kampanya tanımlandı."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "deha_token": []
                    }
                ]
            }
        },
        "/api/seller/offer/product/update": {
            "post": {
                "tags": [
                    "Kampanya"
                ],
                "summary": "Kategori Kampanya Güncelle",
                "description": "Var olan üründeki var olan bir kampanyayı günceller.",
                "operationId": "offerProductUpdate",
                "requestBody": {
                    "description": "Var olan üründeki var olan bir kampanyayı günceller.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "discount_id"
                                ],
                                "properties": {
                                    "discount_id": {
                                        "type": "integer",
                                        "example": "Kampanya Id"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": "Ürün Id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kampanya tanımlandı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kampanya tanımlandı."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "deha_token": []
                    }
                ]
            }
        },
...
        "/api/seller/offer/product/delete": {
            "post": {
                "tags": [
                    "Kampanya"
                ],
                "summary": "Ürün Kampanya Sil",
                "description": "Var olan üründeki var olan bir kampanyayı siler.",
                "operationId": "offerProductDelete",
                "requestBody": {
                    "description": "Var olan üründeki var olan bir kampanyayı siler.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "offer_id"
                                ],
                                "properties": {
                                    "offer_id": {
                                        "type": "integer",
                                        "example": "Kampanya Tanım Id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kampanya silindi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kampanya silindi."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "deha_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "deha_token": {
                "type": "apiKey",
                "description": "Login işleminden dönen tokenType ve tokeni sırasıyla boşluk bırakarak giriniz.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}